dW edit of lab
I made many of the same changes I made to Lab 3: 
Reworded some of the steps to follow dW editorial style guidelines ("click" vs. "click on") and formatting (highlighting or not highlighting page/code elements). 
I did not alter the order of the steps so that I would not affect the overall user code and placeholder numbers.